Extract dialog out
